This is my solution to task0 for Log Message flow visualization.
To observe the results along with the codes please follow these steps:
1. compile and install syslog-ng with the 4 modified file (logpipe.c,
logpipe.h, cfg-tree.c, cfg-tree.h)
2. fill a config file with named (!) elements or just simply copy the
attached syslog-ng.conf, which is prepared with defined log pipelines.
3. start syslog-ng
4. add read and write access permissions to the created
/var/log/logcap.log ( I couldn't find a suitable solution yet to evade this
step )
5. make some logging events ( if you copied syslog-ng.conf you should
create /var/log/source.log file and write into it )
6. start the GUI (with LogFlowGUIapp attached)
7. make more logging events

When the GUI is running after the above mentioned steps, you can choose
an event from the list on the left of the gui (UNIX times) to get a basic
animation of the message flow.

This is just schematic plan for the project, there are some temporary
solutions in it, but it reflects my ideas about the setup.

Signed-off-by Kristof Havasi <kristof.havasi95@gmail.com>